Update tests to remove uses of deprecated methods
A bunch of tests are now redundant or meaningless and were removed. The
remaining tests were reorganized and improved.

The migration tests have been split into separate tests for if a migration is
needed, if the migration is called at the correct time, and if the actual
migration works correctly. When exactly a schema version bump is needed should
now be fully specified by the tests.